+++ src/components/Swap/Summary/Details.tsx
@@ -5,7 +5,7 @@ import { Currency, TradeType } from '@uniswap/sdk-core'
 import Column from 'components/Column'
 import Row from 'components/Row'
 import { Slippage } from 'hooks/useSlippage'
-import { PriceImpact } from 'hooks/useUSDCPriceImpact'
+import { PriceImpact } from 'hooks/usePriceImpact'
 import { useAtomValue } from 'jotai/utils'
 import { useMemo } from 'react'
 import { feeOptionsAtom } from 'state/swap'
