+++ src/components/Swap/Summary/index.tsx
@@ -8,7 +8,7 @@ import { Header } from 'components/Dialog'
 import BaseExpando from 'components/Expando'
 import Row from 'components/Row'
 import { Slippage } from 'hooks/useSlippage'
-import { PriceImpact } from 'hooks/useUSDCPriceImpact'
+import { PriceImpact } from 'hooks/usePriceImpact'
 import { AlertTriangle, BarChart, Info, Spinner } from 'icons'
 import { useCallback, useMemo, useState } from 'react'
 import styled from 'styled-components/macro'
