+++ src/components/Swap/Toolbar/Caption.tsx
@@ -4,7 +4,7 @@ import Column from 'components/Column'
 import Rule from 'components/Rule'
 import Tooltip from 'components/Tooltip'
 import { loadingCss } from 'css/loading'
-import { PriceImpact } from 'hooks/useUSDCPriceImpact'
+import { PriceImpact } from 'hooks/usePriceImpact'
 import { AlertTriangle, Icon, Info, InlineSpinner } from 'icons'
 import { ReactNode, useCallback } from 'react'
 import { InterfaceTrade } from 'state/routing/types'
