+++ src/hooks/swap/useSwapInfo.tsx
@@ -1,14 +1,14 @@
-import { Currency, CurrencyAmount, TradeType } from '@uniswap/sdk-core'
+import { Currency, CurrencyAmount, Token, TradeType } from '@uniswap/sdk-core'
 import { useWeb3React } from '@web3-react/core'
 import { INVALID_TRADE, useRouterTrade } from 'hooks/routing/useRouterTrade'
 import { useCurrencyBalances } from 'hooks/useCurrencyBalance'
 import useSlippage, { DEFAULT_SLIPPAGE, Slippage } from 'hooks/useSlippage'
-import useUSDCPriceImpact, { PriceImpact, toHumanReadablePercent } from 'hooks/useUSDCPriceImpact'
+import { useUSDCValue } from 'hooks/useUSDCPrice'
+import { PriceImpact, usePriceImpact } from 'hooks/usePriceImpact'
 import { useAtomValue } from 'jotai/utils'
 import { createContext, PropsWithChildren, useContext, useMemo } from 'react'
 import { InterfaceTrade, TradeState } from 'state/routing/types'
 import { Field, swapAtom } from 'state/swap'
-import { computeRealizedPriceImpact, getPriceImpactWarning, largerPercentValue } from 'utils/prices'
 import tryParseCurrencyAmount from 'utils/tryParseCurrencyAmount'
 
 import { useIsWrap } from './useWrapCallback'
@@ -65,26 +65,10 @@ function useComputeSwapInfo(routerUrl?: string): SwapInfo {
   )
 
   const slippage = useSlippage(trade.trade)
-  const {
-    inputUSDC,
-    outputUSDC,
-    impact: stablecoinPriceImpact,
-  } = useUSDCPriceImpact(trade.trade?.inputAmount, trade.trade?.outputAmount)
+  const inputUSDCValue = useUSDCValue(trade.trade?.inputAmount)
+  const outputUSDCValue = useUSDCValue(trade.trade?.outputAmount)
 
-  const impact: PriceImpact | undefined = useMemo(() => {
-    const marketPriceImpact = trade.trade ? computeRealizedPriceImpact(trade.trade) : undefined
-    if (!stablecoinPriceImpact && !marketPriceImpact) {
-      return undefined
-    }
-    const percent = largerPercentValue(marketPriceImpact, stablecoinPriceImpact?.percent)
-    return percent
-      ? {
-          percent,
-          warning: getPriceImpactWarning(percent),
-          toString: () => toHumanReadablePercent(percent),
-        }
-      : undefined
-  }, [stablecoinPriceImpact, trade])
+  const impact = usePriceImpact(trade.trade, { inputUSDCValue, outputUSDCValue })
 
   return useMemo(
     () => ({
@@ -92,13 +76,13 @@ function useComputeSwapInfo(routerUrl?: string): SwapInfo {
         currency: currencyIn,
         amount: amountIn,
         balance: balanceIn,
-        usdc: inputUSDC,
+        usdc: inputUSDCValue,
       },
       [Field.OUTPUT]: {
         currency: currencyOut,
         amount: amountOut,
         balance: balanceOut,
-        usdc: outputUSDC,
+        usdc: outputUSDCValue,
       },
       trade,
       slippage,
@@ -112,8 +96,8 @@ function useComputeSwapInfo(routerUrl?: string): SwapInfo {
       currencyIn,
       currencyOut,
       impact,
-      inputUSDC,
-      outputUSDC,
+      inputUSDCValue,
+      outputUSDCValue,
       slippage,
       trade,
     ]
