+++ src/hooks/usePriceImpact.ts
@@ -1,0 +1,42 @@
+import { Currency, CurrencyAmount, Percent, Token, TradeType } from '@uniswap/sdk-core'
+import { useUSDCValue } from 'hooks/useUSDCPrice'
+import { useMemo } from 'react'
+import { InterfaceTrade } from 'state/routing/types'
+import { computeFiatValuePriceImpact } from 'utils/computeFiatValuePriceImpact'
+import { computeRealizedPriceImpact, getPriceImpactWarning, largerPercentValue } from 'utils/prices'
+
+export interface PriceImpact {
+  percent: Percent
+  warning?: 'warning' | 'error'
+  toString(): string
+}
+
+export function usePriceImpact(
+  trade: InterfaceTrade<Currency, Currency, TradeType> | undefined,
+  {
+    inputUSDCValue,
+    outputUSDCValue,
+  }: { inputUSDCValue: CurrencyAmount<Token> | undefined; outputUSDCValue: CurrencyAmount<Token> | undefined }
+) {
+  return useMemo(() => {
+    const fiatPriceImpact = computeFiatValuePriceImpact(inputUSDCValue, outputUSDCValue)
+    const marketPriceImpact = trade ? computeRealizedPriceImpact(trade) : undefined
+    if (!fiatPriceImpact && !marketPriceImpact) {
+      return undefined
+    }
+    const percent = largerPercentValue(marketPriceImpact, fiatPriceImpact)
+    return percent
+      ? {
+          percent,
+          warning: getPriceImpactWarning(percent),
+          toString: () => toHumanReadablePercent(percent),
+        }
+      : undefined
+  }, [trade])
+}
+
+export function toHumanReadablePercent(priceImpact: Percent): string {
+  const sign = priceImpact.lessThan(0) ? '+' : ''
+  const number = parseFloat(priceImpact.multiply(-1)?.toSignificant(3))
+  return `${sign}${number}%`
+}
